+++ backend/canisters/community/api/src/updates/c2c_notify_events.rs
@@ -1,0 +1,12 @@
+use crate::LocalGroupIndexEvent;
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct Args {
+    pub events: Vec<LocalGroupIndexEvent>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub enum Response {
+    Success,
+}