+++ backend/canisters/community/impl/src/lifecycle/post_upgrade.rs
@@ -18,10 +18,12 @@ fn post_upgrade(args: Args) {
     let memory = get_upgrades_memory();
     let reader = get_reader(&memory);
 
-    let (data, errors, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>, Vec<LogEntry>) =
+    let (mut data, errors, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>, Vec<LogEntry>) =
         msgpack::deserialize(reader).unwrap();
 
     canister_logger::init_with_logs(data.test_mode, errors, logs, traces);
+
+    data.members.remove_dangling_member_channel_links();
 
     let env = init_env(data.rng_seed);
     init_state(env, data, args.wasm_version);
@@ -38,6 +40,6 @@ fn post_upgrade(args: Args) {
         let now = state.env.now();
         state
             .data
-            .record_instructions_count(InstructionCountFunctionId::PostUpgrade, now)
+            .record_instructions_count(InstructionCountFunctionId::PostUpgrade, now);
     });
 }