+++ backend/canisters/community/impl/src/model/members.rs
@@ -8,7 +8,8 @@ use rand::RngCore;
 use serde::{Deserialize, Serialize};
 use stable_memory_map::StableMemoryMap;
 use std::collections::btree_map::Entry::Vacant;
-use std::collections::{BTreeMap, BTreeSet};
+use std::collections::{BTreeMap, BTreeSet, HashSet};
+use tracing::info;
 use types::{
     is_default, ChannelId, CommunityMember, CommunityPermissions, CommunityRole, PushIfNotContains, TimestampMillis,
     Timestamped, UserId, UserType, Version,
@@ -41,6 +42,34 @@ pub struct CommunityMembers {
 }
 
 impl CommunityMembers {
+    pub fn remove_dangling_member_channel_links(&mut self) {
+        let user_ids: HashSet<_> = self.members_map.user_ids().into_iter().collect();
+
+        let to_remove: Vec<_> = self
+            .members_and_channels
+            .keys()
+            .filter(|u| !user_ids.contains(u))
+            .copied()
+            .collect();
+
+        for user_id in to_remove {
+            let channel_count = self.members_and_channels.remove(&user_id).unwrap().len();
+            info!(%user_id, channel_count, "Removed dangling member channel links");
+        }
+
+        let to_remove: Vec<_> = self
+            .member_channel_links_removed
+            .keys()
+            .filter(|(u, _)| !user_ids.contains(u))
+            .copied()
+            .collect();
+
+        for (user_id, channel_id) in to_remove {
+            self.member_channel_links_removed.remove(&(user_id, channel_id));
+            info!(%user_id, "Removed dangling member channel link removed");
+        }
+    }
+
     pub fn new(
         creator_principal: Principal,
         creator_user_id: UserId,
@@ -691,10 +720,7 @@ pub struct CommunityMemberInternal {
 
 impl CommunityMemberInternal {
     pub fn accept_rules(&mut self, version: Version, now: TimestampMillis) -> bool {
-        let already_accepted = self
-            .rules_accepted
-            .as_ref()
-            .map_or(false, |accepted| version <= accepted.value);
+        let already_accepted = self.rules_accepted.as_ref().is_some_and(|accepted| version <= accepted.value);
 
         if !already_accepted {
             self.rules_accepted = Some(Timestamped::new(version, now));