+++ backend/canisters/group/impl/src/updates/accept_p2p_swap.rs
@@ -51,7 +51,7 @@ async fn accept_p2p_swap(args: Args) -> Response {
                         .chat
                         .members
                         .get(&message.sender)
-                        .map_or(false, |m| !m.user_type().is_bot())
+                        .is_some_and(|m| !m.user_type().is_bot())
                     {
                         state.data.user_event_sync_queue.push(
                             message.sender,