+++ backend/canisters/group_index/impl/src/model/mod.rs
@@ -1,6 +1,7 @@
 pub mod cached_hot_groups;
 pub mod deleted_communities;
 pub mod deleted_groups;
+pub mod local_group_index_event_batch;
 pub mod local_group_index_map;
 pub mod moderation_flags;
 pub mod private_communities;