+++ backend/canisters/group_index/impl/src/model/public_group_and_community_names.rs
@@ -9,6 +9,18 @@ pub struct PublicGroupAndCommunityNames {
 }
 
 impl PublicGroupAndCommunityNames {
+    pub fn check(&self, name: &str) -> CheckNameResult {
+        if self.reserved.contains_key(name) {
+            return CheckNameResult::Reserved;
+        }
+
+        if let Some(canister_id) = self.names.get(name) {
+            return CheckNameResult::Taken(*canister_id);
+        }
+
+        CheckNameResult::Available
+    }
+
     pub fn is_name_taken(&self, name: &str) -> bool {
         self.names.contains_key(name) || self.reserved.contains_key(name)
     }
@@ -20,8 +32,17 @@ impl PublicGroupAndCommunityNames {
 
     // Only removes the entry if both the name and canister_id match
     pub fn remove(&mut self, name: &str, canister_id: CanisterId) -> bool {
-        if self.names.get(name).map_or(false, |c| *c == canister_id) {
+        if self.names.get(name).is_some_and(|c| *c == canister_id) {
             self.names.remove(name).is_some()
+        } else {
+            false
+        }
+    }
+
+    pub fn rename(&mut self, curr_name: &str, new_name: &str, canister_id: CanisterId) -> bool {
+        if self.remove(curr_name, canister_id) {
+            self.insert(new_name, canister_id);
+            true
         } else {
             false
         }
@@ -39,4 +60,10 @@ impl PublicGroupAndCommunityNames {
     pub fn unreserve_name(&mut self, name: &str) -> bool {
         self.reserved.remove(name).is_some()
     }
+}
+
+pub enum CheckNameResult {
+    Available,
+    Reserved,
+    Taken(CanisterId),
 }