+++ backend/canisters/user_index/impl/src/lib.rs
@@ -145,7 +145,7 @@ impl RuntimeState {
 
     pub fn push_event_to_all_local_user_indexes(&mut self, event: LocalUserIndexEvent, except: Option<CanisterId>) {
         for canister_id in self.data.local_index_map.canisters() {
-            if except.map_or(true, |id| id != *canister_id) {
+            if except != Some(*canister_id) {
                 self.data.user_index_event_sync_queue.push(*canister_id, event.clone());
             }
         }